PRO-385: подбор профессий по Pearson по полному 6-мерному RIASEC
Заменяет нестабильный 3-буквенный код на корреляцию с O*NET-векторами; для трёх профессий без SOC оставляем code-fallback.

+++ app/services/student_context.py
@@ -71,7 +71,7 @@ def _context_careers(analysis: AnalysisResult | None) -> list[ContextCareer]:
                 slug=c.get("slug", ""),
                 name=c.get("name", ""),
                 holland_code=c.get("holland_code", ""),
-                match_score=int(c.get("match_score", 0)),
+                match_score=float(c.get("match_score", 0)),
                 description=c.get("description", ""),
                 professions=list(c.get("professions", [])),
                 skills_needed=list(c.get("skills_needed", [])),
